prototipo menu, com integrações de tudo; está meio bugado ainda

+++ fileModule/fileModule.h
@@ -1,9 +1,10 @@
 #pragma once
 #include <string>
+#include "../include/trie.h"
 namespace trie {
 struct TrieNode;
 }
 namespace fileModule
 {
-  void processNodesToLabel(trie::TrieNode *trieRoot, void (*cb)(trie::TrieNode *trieRoot, std::string word, unsigned long nodeId));
+  void processNodesToLabel(trie::TrieNode *trieRoot, void (*cb)(trie::TrieNode *trieRoot, std::string word, unsigned long long nodeId));
 }
